[testing] unified functions to start scheduler in testing
Implemented function "run_scheduler" as mixin which can be used in
the class ActionDispatch::IntegrationTest and script/start_test_backend
