+++ src/api/test/test_helper.rb
@@ -3,6 +3,7 @@ ENV['RAILS_ENV'] = 'test'
 require 'simplecov'
 require 'coveralls'
 require "minitest/reporters"
+
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 if ENV['DO_COVERAGE']
@@ -371,6 +372,8 @@ end
 module ActionDispatch
   class IntegrationTest
 
+    include TestBackendTasks
+
     def teardown
       Rails.cache.clear
       reset_auth
@@ -434,35 +437,8 @@ module ActionDispatch
       super(hash)
     end
 
-    def run_dispatcher
-      Rails.logger.debug 'run dispatcher'
-      perlopts="-I#{Rails.root}/../backend -I#{Rails.root}/../backend/build"
-      IO.popen("cd #{backend_config}; exec perl #{perlopts} ./bs_dispatch --test-mode") do |io|
-        # just for waiting until dispatcher finishes
-        io.each { |line| Rails.logger.debug("dispatcher: #{line.strip.chomp}") unless line.blank? }
-      end
-    end
-
-    def run_publisher
-      Rails.logger.debug 'run publisher'
-      perlopts="-I#{Rails.root}/../backend -I#{Rails.root}/../backend/build"
-      IO.popen("cd #{backend_config}; exec perl #{perlopts} ./bs_publish --testmode") do |io|
-        # just for waiting until publisher finishes
-        io.each { |line| Rails.logger.debug("publisher: #{line.strip.chomp}") unless line.blank? }
-      end
-    end
-
     def wait_for_scheduler_start
       Suse::Backend.wait_for_scheduler_start
-    end
-
-    def run_scheduler(arch)
-      Rails.logger.debug "RUN_SCHEDULER #{arch}"
-      perlopts="-I#{Rails.root}/../backend -I#{Rails.root}/../backend/build"
-      IO.popen("cd #{backend_config}; exec perl #{perlopts} ./bs_sched --testmode #{arch}") do |io|
-        # just for waiting until scheduler finishes
-        io.each { |line| Rails.logger.debug("scheduler(#{arch}): #{line.strip.chomp}") unless line.blank? }
-      end
     end
 
     def login_king
